+++ models/register.go
@@ -3,7 +3,7 @@ package models
 import "time"
 
 type Register struct {
-	RegId       uint32    `gorm:"primary_key;auto_increment" json:"reg_id"`
+	RegId       uint64    `gorm:"primary_key;auto_increment" json:"reg_id"`
 	RegName     string    `gorm:"size:30" json:"reg_name"`
 	RegCompany  string    `gorm:"size:30" json:"reg_company"`
 	RegCheckIn  time.Time `json:"reg_check_in"`
